Move toolbar buttons to their own files
The toolbar gets wired with its buttons during the
creation of the extended Prism object.
